apps/asap-server/test/handlers/webhooks/webhook-user.test.ts
import { User, WebhookPayload } from '@asap-hub/squidex';
import { APIGatewayProxyResult } from 'aws-lambda';
import { EventBridge } from 'aws-sdk';
import { userWebhookFactory } from '../../../src/handlers/webhooks/webhook-user';
import { signPayload } from '../../../src/utils/validate-squidex-request';
import { userPublishedEvent } from '../../fixtures/users.fixtures';
import { getApiGatewayEvent } from '../../helpers/events';

describe('User webhook', () => {
  const evenBridgeMock = {
    putEvents: jest.fn().mockReturnValue({ promise: jest.fn() }),
  } as unknown as jest.Mocked<EventBridge>;
  const handler = userWebhookFactory(evenBridgeMock);

  afterEach(() => {
    jest.clearAllMocks();
  });

  test('Should return 403 when the request is not signed correctly', async () => {
    const event = getApiGatewayEvent({
      headers: {
        'x-signature': 'XYZ',
      },
      body: JSON.stringify(userPublishedEvent),
    });

    const res = (await handler(event)) as APIGatewayProxyResult;

    expect(res.statusCode).toStrictEqual(403);
    expect(evenBridgeMock.putEvents).not.toHaveBeenCalled();
  });

  test('Should return 204 and not raise an event when the event type is not supported', async () => {
    const res = (await handler(
      createSignedPayload({
        ...userPublishedEvent,
        type: 'UsersUpdated',
      }),
    )) as APIGatewayProxyResult;

    expect(res.statusCode).toStrictEqual(204);
    expect(evenBridgeMock.putEvents).not.toHaveBeenCalled();
  });

  test('Should return 200 and put an event into the event bus', async () => {
    const res = (await handler(
      createSignedPayload(userPublishedEvent),
    )) as APIGatewayProxyResult;

    expect(res.statusCode).toStrictEqual(200);
    expect(evenBridgeMock.putEvents).toHaveBeenCalled();
  });
});

const createSignedPayload = (payload: WebhookPayload<User>) =>
  getApiGatewayEvent({
    headers: {
      'x-signature': signPayload(payload),
    },
    body: JSON.stringify(payload),
  });
